Fixed Drawing

+++ GlobalVariables.h
@@ -13,7 +13,7 @@ static int dist_obs = 600;
 static double zoom = 1;
 static double default_angle = 0.01; //unghiul cu care se roteste corpul la cerinta utilizatorului
 static int draw_precision = 1; //precizia cu care se deseneaza. poate fi 1/5/10 aproximandu-se coordonatele mouse-ului la un multiplu al acesteia
-static double offsetX = 0, offsetY = 0; //modifica pozitia corpurilor in scena
+static int offsetX = 0, offsetY = 0; //modifica pozitia corpurilor in scena
 static int last_clicked_button = 0; //retine indicele ultimului buton apasat din meniul din stanga
 static int vp_tl_x = 40, vp_tl_y = 30; // coordonate implicite stanga sus viewport
 static int vp_dr_x = 1080, vp_dr_y = 720; // coordonate implicite dreapta jos viewport
